services/agent-api/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
pydantic==2.10.*
pydantic-settings==2.7.*
redis[hiredis]==5.2.*
httpx==0.28.*
openai==1.58.*
anthropic==0.40.*
gitpython==3.1.*
python-slugify==8.0.*
python-multipart==0.0.*
jinja2==3.1.*
json-repair==0.30.*
tenacity==9.0.*
google-api-python-client==2.150.*
google-auth==2.36.*
google-auth-httplib2==0.2.*
slowapi==0.1.*
loguru==0.7.*
beautifulsoup4==4.12.*
